reorganizing the global params stuff
