+++ packages/javascript/bh-shared-ui/src/components/Navigation/AppNavigate.test.tsx
@@ -15,7 +15,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { Route, Routes } from 'react-router-dom';
-import { GloballySupportedSearchParams } from '../..';
+
+import { GloballySupportedSearchParams } from '../../hooks/useGlobalParamsExtension';
 import { render } from '../../test-utils';
 import { AppNavigate } from './AppNavigate';
