+++ packages/javascript/bh-shared-ui/src/utils/searchParams/searchParams.ts
@@ -15,15 +15,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { createSearchParams, NavigateOptions, Path, To, useSearchParams } from 'react-router-dom';
-import { EnvironmentQueryParams } from '../../hooks/useEnvironmentParams';
-import { ExploreQueryParams } from '../../hooks/useExploreParams';
 
 // FUTURE DEV: SetURLSearchParams is in both v6 and v7, but v6 uses this type internally and v7 exports it.
 // When we upgrade to v7, we can import SetURLSearchParams from react-router-dom
 type SetURLSearchParams = ReturnType<typeof useSearchParams>[1];
-
-export type SearchParamKeys = keyof EnvironmentQueryParams | keyof ExploreQueryParams;
-export const GloballySupportedSearchParams = ['environmentId', 'environmentAggregation'] satisfies SearchParamKeys[];
 
 type EmptyParam = undefined | null | '';
